server/requirements.txt
# Django and Django REST Framework
Django==5.0.6
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1

# Database adapters
psycopg2-binary==2.9.9  # For PostgreSQL
dj-database-url==2.1.0  # For database URL parsing

# Authentication and security
django-allauth==0.57.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1
django-rest-knox==4.2.0

# API documentation
drf-yasg==1.21.7
pyyaml==6.0.1
uritemplate==4.1.1

# File handling and processing
Pillow==10.2.0
reportlab==4.0.8
python-magic==0.4.27
xhtml2pdf==0.2.13
openpyxl==3.1.2
python-docx==1.1.0
pypdf2==3.0.1

# Utilities
python-dateutil==2.9.0
pytz==2024.1
six==1.16.0
requests==2.31.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

# Environment management
python-dotenv==1.0.1
django-environ==0.11.2

# AWS S3 support (if using S3 for file storage)
boto3==1.34.51
django-storages==1.14.2

# Testing
pytest==7.4.4
pytest-django==4.8.0
factory-boy==3.3.0
faker==23.2.1

# Debugging and development tools
django-debug-toolbar==4.3.0
ipython==8.21.0
django-extensions==3.2.3